remove deprecated resources (nocompat)
