--- docs/resources/cluster_background_persistence.md
@@ -1,127 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "chalk_cluster_background_persistence Resource - chalk"
-subcategory: ""
-description: |-
-  ~> Deprecated Use chalk_unmanaged_cluster_background_persistence instead.
-  Chalk cluster background persistence resource
-  Required permissions: team.admin, deploy.read
----
-
-# chalk_cluster_background_persistence (Resource)
-
-~> **Deprecated** Use `chalk_unmanaged_cluster_background_persistence` instead.
-
-Chalk cluster background persistence resource
-
-**Required permissions:** `team.admin`, `deploy.read`
-
-
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `kube_cluster_id` (String) Kubernetes cluster ID
-- `namespace` (String) Kubernetes namespace
-- `service_account_name` (String) Service account name
-- `writers` (Attributes List) Background persistence writers (see [below for nested schema](#nestedatt--writers))
-
-### Optional
-
-- `api_server_host` (String) API server host
-- `bigquery_parquet_upload_subscription_id` (String) BigQuery parquet upload subscription ID
-- `bigquery_streaming_write_subscription_id` (String) BigQuery streaming write subscription ID
-- `bigquery_streaming_write_topic` (String) BigQuery streaming write topic
-- `bq_upload_bucket` (String) BigQuery upload bucket
-- `bq_upload_topic` (String) BigQuery upload topic
-- `bus_writer_image_bswl` (String) BSWL bus writer image
-- `bus_writer_image_go` (String) Go bus writer image
-- `bus_writer_image_python` (String) Python bus writer image
-- `bus_writer_image_rust` (String) Rust bus writer image
-- `google_cloud_project` (String) Google Cloud project
-- `kafka_bootstrap_servers` (String) Kafka bootstrap servers
-- `kafka_dlq_topic` (String) Kafka DLQ topic
-- `kafka_sasl_secret` (String) Kafka SASL secret
-- `metrics_bus_subscription_id` (String) Metrics bus subscription ID
-- `metrics_bus_topic_id` (String) Metrics bus topic ID
-- `operation_subscription_id` (String) Operation subscription ID
-- `query_log_result_topic` (String) Query log result topic
-- `query_log_subscription_id` (String) Query log subscription ID
-- `redis_lightning_supports_has_many` (Boolean) Redis lightning supports has many
-- `result_bus_offline_store_subscription_id` (String) Result bus offline store subscription ID
-- `result_bus_online_store_subscription_id` (String) Result bus online store subscription ID
-- `result_bus_topic_id` (String) Result bus topic ID
-- `snowflake_storage_integration_name` (String) Snowflake storage integration name
-
-### Read-Only
-
-- `id` (String) Background persistence identifier
-
-<a id="nestedatt--writers"></a>
-### Nested Schema for `writers`
-
-Required:
-
-- `bus_subscriber_type` (String) Bus subscriber type
-- `name` (String) Writer name
-
-Optional:
-
-- `additional_env_vars` (Map of String) Additional environment variables to set for the writer
-- `default_replica_count` (Number) Default replica count
-- `gke_spot` (Boolean) GKE spot instances
-- `hpa_specs` (Attributes) HPA specifications (see [below for nested schema](#nestedatt--writers--hpa_specs))
-- `image_override` (String) Image override
-- `kafka_consumer_group_override` (String) Kafka consumer group override
-- `limit` (Attributes) Resource limits (see [below for nested schema](#nestedatt--writers--limit))
-- `load_writer_configmap` (Boolean) Load writer configmap
-- `max_batch_size` (Number) Maximum batch size
-- `message_processing_concurrency` (Number) Message processing concurrency
-- `metadata_sql_ssl_ca_cert_secret` (String) Metadata SQL SSL CA cert secret
-- `metadata_sql_ssl_client_cert_secret` (String) Metadata SQL SSL client cert secret
-- `metadata_sql_ssl_client_key_secret` (String) Metadata SQL SSL client key secret
-- `metadata_sql_uri_secret` (String) Metadata SQL URI secret
-- `nodepool` (String) Nodepool to pin the writer to. On GCP this becomes a GKE node selector plus toleration; on AWS it maps to the Karpenter nodepool. When unset, the writer runs on the default nodepool.
-- `offline_store_inserter_db_type` (String) Offline store inserter DB type
-- `query_table_write_drop_ratio` (String) Query table write drop ratio
-- `request` (Attributes) Resource requests (see [below for nested schema](#nestedatt--writers--request))
-- `results_writer_skip_producing_feature_metrics` (Boolean) Results writer skip producing feature metrics
-- `storage_cache_prefix` (String) Storage cache prefix
-- `version` (String) Writer version
-
-<a id="nestedatt--writers--hpa_specs"></a>
-### Nested Schema for `writers.hpa_specs`
-
-Required:
-
-- `hpa_pubsub_subscription_id` (String) HPA pubsub subscription ID
-
-Optional:
-
-- `hpa_max_replicas` (Number) HPA maximum replicas
-- `hpa_min_replicas` (Number) HPA minimum replicas
-- `hpa_target_average_value` (Number) HPA target average value
-
-
-<a id="nestedatt--writers--limit"></a>
-### Nested Schema for `writers.limit`
-
-Optional:
-
-- `cpu` (String) CPU resource specification
-- `ephemeral_storage` (String) Ephemeral storage resource specification
-- `memory` (String) Memory resource specification
-- `storage` (String) Storage resource specification
-
-
-<a id="nestedatt--writers--request"></a>
-### Nested Schema for `writers.request`
-
-Optional:
-
-- `cpu` (String) CPU resource specification
-- `ephemeral_storage` (String) Ephemeral storage resource specification
-- `memory` (String) Memory resource specification
-- `storage` (String) Storage resource specification
