--- docs/resources/environment.md
@@ -1,46 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "chalk_environment Resource - chalk"
-subcategory: ""
-description: |-
-  ~> Deprecated Use chalk_unmanaged_environment or chalk_managed_environment instead.
-  Chalk environment resource
-  Required permissions: project.create, deploy.create, deploy.read, environment.create (team-scoped)
----
-
-# chalk_environment (Resource)
-
-~> **Deprecated** Use `chalk_unmanaged_environment` or `chalk_managed_environment` instead.
-
-Chalk environment resource
-
-**Required permissions:** `project.create`, `deploy.create`, `deploy.read`, `environment.create` *(team-scoped)*
-
-
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `id` (String) Environment identifier
-- `name` (String) Environment name
-- `project_id` (String) Project ID
-
-### Optional
-
-- `additional_env_vars` (Map of String) Additional environment variables
-- `branch_url` (String) Branch URL
-- `engine_docker_registry_path` (String) Engine Docker registry path
-- `feature_store_secret` (String) Feature store secret reference
-- `kube_cluster_id` (String) Kubernetes cluster ID
-- `kube_job_namespace` (String) Kubernetes job namespace
-- `kube_service_account_name` (String) Kubernetes service account name
-- `managed` (Boolean) Whether to bootstrap cloud infrastructure
-- `online_store_kind` (String) Online store kind
-- `online_store_secret` (String) Online store secret reference
-- `private_pip_repositories` (String) Private pip repositories
-- `service_url` (String) Service URL
-- `source_bundle_bucket` (String) Source bundle bucket
-- `specs_config_json` (String) Specs config JSON
-- `worker_url` (String) Worker URL
