--- internal/provider/environment_resource.go
@@ -1,657 +1,0 @@
-// Deprecated: This resource is deprecated. You likely want to modify [un]managed_environment_resource.go instead.
-
-package provider
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/chalk-ai/terraform-provider-chalk/client"
-
-	"connectrpc.com/connect"
-	serverv1 "github.com/chalk-ai/chalk-go/gen/chalk/server/v1"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"google.golang.org/protobuf/types/known/fieldmaskpb"
-)
-
-var _ resource.Resource = &EnvironmentResource{}
-var _ resource.ResourceWithImportState = &EnvironmentResource{}
-
-func NewEnvironmentResource() resource.Resource {
-	return &EnvironmentResource{}
-}
-
-type EnvironmentResource struct {
-	client *client.Manager
-}
-
-type EnvironmentResourceModel struct {
-	// Should not be settable IDK, probably should have a random name
-	Id        types.String `tfsdk:"id"`
-	Name      types.String `tfsdk:"name"`
-	ProjectId types.String `tfsdk:"project_id"`
-
-	KubeClusterId            types.String `tfsdk:"kube_cluster_id"`
-	EngineDockerRegistryPath types.String `tfsdk:"engine_docker_registry_path"`
-	OnlineStoreSecret        types.String `tfsdk:"online_store_secret"`
-	OnlineStoreKind          types.String `tfsdk:"online_store_kind"`
-	FeatureStoreSecret       types.String `tfsdk:"feature_store_secret"`
-	SpecsConfigJson          types.String `tfsdk:"specs_config_json"`
-	Managed                  types.Bool   `tfsdk:"managed"`
-
-	//TODO remove, deprecated
-	SourceBundleBucket types.String `tfsdk:"source_bundle_bucket"`
-	//TODO fallback from gateway
-	ServiceUrl types.String `tfsdk:"service_url"`
-	//TODO fallback from gateway
-	WorkerUrl types.String `tfsdk:"worker_url"`
-	//TODO fallback from gateway
-	BranchUrl types.String `tfsdk:"branch_url"`
-	//TODO template if not provided
-	KubeJobNamespace types.String `tfsdk:"kube_job_namespace"`
-	//TODO template if not provided
-	KubeServiceAccountName types.String `tfsdk:"kube_service_account_name"`
-
-	//Optional Fields
-	PrivatePipRepositories types.String `tfsdk:"private_pip_repositories"`
-	AdditionalEnvVars      types.Map    `tfsdk:"additional_env_vars"`
-}
-
-func (r *EnvironmentResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_environment"
-}
-
-func (r *EnvironmentResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-		DeprecationMessage:  "Use chalk_unmanaged_environment or chalk_managed_environment instead.",
-		MarkdownDescription: "~> **Deprecated** Use `chalk_unmanaged_environment` or `chalk_managed_environment` instead.\n\nChalk environment resource",
-
-		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				MarkdownDescription: "Environment identifier",
-				Required:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"name": schema.StringAttribute{
-				MarkdownDescription: "Environment name",
-				Required:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"project_id": schema.StringAttribute{
-				MarkdownDescription: "Project ID",
-				Required:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"source_bundle_bucket": schema.StringAttribute{
-				MarkdownDescription: "Source bundle bucket",
-				Optional:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"kube_cluster_id": schema.StringAttribute{
-				MarkdownDescription: "Kubernetes cluster ID",
-				Optional:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"online_store_secret": schema.StringAttribute{
-				MarkdownDescription: "Online store secret reference",
-				Optional:            true,
-			},
-			"online_store_kind": schema.StringAttribute{
-				MarkdownDescription: "Online store kind",
-				Optional:            true,
-			},
-			"feature_store_secret": schema.StringAttribute{
-				MarkdownDescription: "Feature store secret reference",
-				Optional:            true,
-			},
-			"private_pip_repositories": schema.StringAttribute{
-				MarkdownDescription: "Private pip repositories",
-				Optional:            true,
-			},
-			"additional_env_vars": schema.MapAttribute{
-				MarkdownDescription: "Additional environment variables",
-				Optional:            true,
-				ElementType:         types.StringType,
-			},
-			"specs_config_json": schema.StringAttribute{
-				MarkdownDescription: "Specs config JSON",
-				Optional:            true,
-				Computed:            true,
-			},
-			"service_url": schema.StringAttribute{
-				MarkdownDescription: "Service URL",
-				Optional:            true,
-			},
-			"worker_url": schema.StringAttribute{
-				MarkdownDescription: "Worker URL",
-				Optional:            true,
-			},
-			"branch_url": schema.StringAttribute{
-				MarkdownDescription: "Branch URL",
-				Optional:            true,
-			},
-			"kube_job_namespace": schema.StringAttribute{
-				MarkdownDescription: "Kubernetes job namespace",
-				Optional:            true,
-			},
-			"kube_service_account_name": schema.StringAttribute{
-				MarkdownDescription: "Kubernetes service account name",
-				Optional:            true,
-			},
-			"engine_docker_registry_path": schema.StringAttribute{
-				MarkdownDescription: "Engine Docker registry path",
-				Optional:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"managed": schema.BoolAttribute{
-				MarkdownDescription: "Whether to bootstrap cloud infrastructure",
-				Optional:            true,
-			},
-		},
-	}
-}
-
-func (r *EnvironmentResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
-	if req.ProviderData == nil {
-		return
-	}
-
-	client, ok := req.ProviderData.(*client.Manager)
-
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected *client.Manager, got: %T. Please report this issue to the provider developers.", req.ProviderData),
-		)
-
-		return
-	}
-
-	r.client = client
-}
-
-func (r *EnvironmentResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var data EnvironmentResourceModel
-
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Get team client from client.Manager
-	tc := r.client.NewTeamClient(ctx)
-
-	createReq := &serverv1.CreateEnvironmentRequest{
-		ProjectId: data.ProjectId.ValueString(),
-		Name:      data.Name.ValueString(),
-	}
-
-	if !data.SourceBundleBucket.IsNull() {
-		createReq.SourceBundleBucket = data.SourceBundleBucket.ValueString()
-	}
-
-	if !data.KubeClusterId.IsNull() {
-		clusterId := data.KubeClusterId.ValueString()
-		createReq.KubeClusterId = &clusterId
-	}
-
-	if !data.EngineDockerRegistryPath.IsNull() {
-		val := data.EngineDockerRegistryPath.ValueString()
-		createReq.EngineDockerRegistryPath = &val
-	}
-
-	// Use the id field as environment_id_override
-	if !data.Id.IsNull() {
-		val := data.Id.ValueString()
-		createReq.EnvironmentIdOverride = &val
-	}
-
-	if !data.Managed.IsNull() {
-		createReq.Managed = data.Managed.ValueBool()
-	}
-
-	_, err := tc.CreateEnvironment(ctx, connect.NewRequest(createReq))
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error Creating Chalk Environment",
-			fmt.Sprintf("Could not create environment: %v", err),
-		)
-		return
-	}
-
-	// If update fields were provided, update the environment
-	if !data.OnlineStoreSecret.IsNull() || !data.FeatureStoreSecret.IsNull() ||
-		!data.PrivatePipRepositories.IsNull() || !data.AdditionalEnvVars.IsNull() ||
-		!data.SpecsConfigJson.IsNull() || !data.ServiceUrl.IsNull() ||
-		!data.WorkerUrl.IsNull() || !data.BranchUrl.IsNull() ||
-		!data.KubeJobNamespace.IsNull() || !data.KubeServiceAccountName.IsNull() ||
-		!data.OnlineStoreKind.IsNull() {
-		updateReq := &serverv1.UpdateEnvironmentRequest{
-			Id:     data.Id.ValueString(),
-			Update: &serverv1.UpdateEnvironmentOperation{},
-		}
-
-		var updateMaskPaths []string
-
-		if !data.OnlineStoreSecret.IsNull() {
-			val := data.OnlineStoreSecret.ValueString()
-			updateReq.Update.OnlineStoreSecret = &val
-			updateMaskPaths = append(updateMaskPaths, "online_store_secret")
-		}
-
-		if !data.OnlineStoreKind.IsNull() {
-			val := data.OnlineStoreKind.ValueString()
-			updateReq.Update.OnlineStoreKind = &val
-			updateMaskPaths = append(updateMaskPaths, "online_store_kind")
-		}
-
-		if !data.FeatureStoreSecret.IsNull() {
-			val := data.FeatureStoreSecret.ValueString()
-			updateReq.Update.FeatureStoreSecret = &val
-			updateMaskPaths = append(updateMaskPaths, "feature_store_secret")
-		}
-
-		if !data.PrivatePipRepositories.IsNull() {
-			val := data.PrivatePipRepositories.ValueString()
-			updateReq.Update.PrivatePipRepositories = &val
-			updateMaskPaths = append(updateMaskPaths, "private_pip_repositories")
-		}
-
-		if !data.AdditionalEnvVars.IsNull() && !data.AdditionalEnvVars.IsUnknown() {
-			envVars := make(map[string]string)
-			diags := data.AdditionalEnvVars.ElementsAs(ctx, &envVars, false)
-			resp.Diagnostics.Append(diags...)
-			if resp.Diagnostics.HasError() {
-				return
-			}
-			updateReq.Update.AdditionalEnvVars = envVars
-			updateMaskPaths = append(updateMaskPaths, "additional_env_vars")
-		}
-
-		if !data.SpecsConfigJson.IsNull() {
-			val := data.SpecsConfigJson.ValueString()
-			updateReq.Update.SpecsConfigJson = &val
-			updateMaskPaths = append(updateMaskPaths, "specs_config_json")
-		}
-
-		if !data.ServiceUrl.IsNull() {
-			val := data.ServiceUrl.ValueString()
-			updateReq.Update.ServiceUrl = &val
-			updateMaskPaths = append(updateMaskPaths, "service_url")
-		}
-
-		if !data.WorkerUrl.IsNull() {
-			val := data.WorkerUrl.ValueString()
-			updateReq.Update.WorkerUrl = &val
-			updateMaskPaths = append(updateMaskPaths, "worker_url")
-		}
-
-		if !data.BranchUrl.IsNull() {
-			val := data.BranchUrl.ValueString()
-			updateReq.Update.BranchUrl = &val
-			updateMaskPaths = append(updateMaskPaths, "branch_url")
-		}
-
-		if !data.KubeJobNamespace.IsNull() {
-			val := data.KubeJobNamespace.ValueString()
-			updateReq.Update.KubeJobNamespace = &val
-			updateMaskPaths = append(updateMaskPaths, "kube_job_namespace")
-		}
-
-		if !data.KubeServiceAccountName.IsNull() {
-			val := data.KubeServiceAccountName.ValueString()
-			updateReq.Update.KubeServiceAccountName = &val
-			updateMaskPaths = append(updateMaskPaths, "kube_service_account_name")
-		}
-
-		updateReq.UpdateMask = &fieldmaskpb.FieldMask{
-			Paths: updateMaskPaths,
-		}
-
-		// Use the environment ID for the header
-		tcUpdate := r.client.NewTeamClient(ctx, data.Id.ValueString())
-
-		_, err = tcUpdate.UpdateEnvironment(ctx, connect.NewRequest(updateReq))
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error Updating Chalk Environment",
-				fmt.Sprintf("Environment was created but could not be updated: %v", err),
-			)
-			return
-		}
-	}
-
-	if data.Managed.ValueBool() {
-		bc := r.client.NewBuilderClient(ctx)
-
-		_, err := bc.CreateEnvironmentCloudResources(ctx, connect.NewRequest(&serverv1.CreateEnvironmentCloudResourcesRequest{
-			EnvironmentId: data.Id.ValueString(),
-		}))
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error Bootstrapping Chalk Environment Cloud Resources",
-				fmt.Sprintf("Environment was created but cloud resources could not be bootstrapped: %v", err),
-			)
-			return
-		}
-	}
-
-	tflog.Trace(ctx, "created a chalk_environment resource")
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-func (r *EnvironmentResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var data EnvironmentResourceModel
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Create auth client first
-	// Create team client with token injection interceptor
-	tc := r.client.NewTeamClient(ctx, data.Id.ValueString())
-
-	env, err := tc.GetEnv(ctx, connect.NewRequest(&serverv1.GetEnvRequest{}))
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error Reading Chalk Environment",
-			fmt.Sprintf("Could not read environment %s: %v", data.Id.ValueString(), err),
-		)
-		return
-	}
-
-	// Update the model with the fetched data
-	e := env.Msg.Environment
-	data.Name = types.StringValue(e.Name)
-	data.ProjectId = types.StringValue(e.ProjectId)
-
-	if e.OnlineStoreSecret != nil {
-		data.OnlineStoreSecret = types.StringValue(*e.OnlineStoreSecret)
-	}
-
-	if e.OnlineStoreKind != nil {
-		data.OnlineStoreKind = types.StringValue(*e.OnlineStoreKind)
-	}
-
-	if e.FeatureStoreSecret != nil {
-		data.FeatureStoreSecret = types.StringValue(*e.FeatureStoreSecret)
-	}
-
-	if e.PrivatePipRepositories != nil {
-		data.PrivatePipRepositories = types.StringValue(*e.PrivatePipRepositories)
-	}
-
-	if e.SourceBundleBucket != nil {
-		data.SourceBundleBucket = types.StringValue(*e.SourceBundleBucket)
-	}
-
-	if e.KubeClusterId != nil {
-		data.KubeClusterId = types.StringValue(*e.KubeClusterId)
-	}
-
-	if len(e.AdditionalEnvVars) > 0 {
-		elements := make(map[string]attr.Value)
-		for k, v := range e.AdditionalEnvVars {
-			elements[k] = types.StringValue(v)
-		}
-		data.AdditionalEnvVars = types.MapValueMust(types.StringType, elements)
-	}
-
-	if e.ServiceUrl != nil {
-		data.ServiceUrl = types.StringValue(*e.ServiceUrl)
-	}
-
-	if e.WorkerUrl != nil {
-		data.WorkerUrl = types.StringValue(*e.WorkerUrl)
-	}
-
-	if e.BranchUrl != nil {
-		data.BranchUrl = types.StringValue(*e.BranchUrl)
-	}
-
-	if e.KubeJobNamespace != nil {
-		data.KubeJobNamespace = types.StringValue(*e.KubeJobNamespace)
-	}
-
-	if e.KubeServiceAccountName != nil {
-		data.KubeServiceAccountName = types.StringValue(*e.KubeServiceAccountName)
-	}
-
-	if e.EngineDockerRegistryPath != nil {
-		data.EngineDockerRegistryPath = types.StringValue(*e.EngineDockerRegistryPath)
-	}
-
-	// Note: managed is a Terraform-only field that controls bootstrap behavior
-	// It is not returned by the server and is preserved from the existing state
-
-	// Note: specs_config_json is not directly available in the Environment message
-	// It is preserved from the existing state during Read operations
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-func (r *EnvironmentResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var data EnvironmentResourceModel
-
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Create auth client first
-	// Create team client with token injection interceptor
-	tc := r.client.NewTeamClient(ctx, data.Id.ValueString())
-
-	updateReq := &serverv1.UpdateEnvironmentRequest{
-		Id:     data.Id.ValueString(),
-		Update: &serverv1.UpdateEnvironmentOperation{},
-	}
-
-	var updateMaskPaths []string
-
-	// Check what fields have changed
-	var state EnvironmentResourceModel
-	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	if !data.OnlineStoreSecret.Equal(state.OnlineStoreSecret) {
-		if !data.OnlineStoreSecret.IsNull() {
-			val := data.OnlineStoreSecret.ValueString()
-			updateReq.Update.OnlineStoreSecret = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "online_store_secret")
-	}
-
-	if !data.OnlineStoreKind.Equal(state.OnlineStoreKind) {
-		if !data.OnlineStoreKind.IsNull() {
-			val := data.OnlineStoreKind.ValueString()
-			updateReq.Update.OnlineStoreKind = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "online_store_kind")
-	}
-
-	if !data.FeatureStoreSecret.Equal(state.FeatureStoreSecret) {
-		if !data.FeatureStoreSecret.IsNull() {
-			val := data.FeatureStoreSecret.ValueString()
-			updateReq.Update.FeatureStoreSecret = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "feature_store_secret")
-	}
-
-	if !data.PrivatePipRepositories.Equal(state.PrivatePipRepositories) {
-		if !data.PrivatePipRepositories.IsNull() {
-			val := data.PrivatePipRepositories.ValueString()
-			updateReq.Update.PrivatePipRepositories = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "private_pip_repositories")
-	}
-
-	if !data.AdditionalEnvVars.Equal(state.AdditionalEnvVars) {
-		if !data.AdditionalEnvVars.IsNull() && !data.AdditionalEnvVars.IsUnknown() {
-			envVars := make(map[string]string)
-			diags := data.AdditionalEnvVars.ElementsAs(ctx, &envVars, false)
-			resp.Diagnostics.Append(diags...)
-			if resp.Diagnostics.HasError() {
-				return
-			}
-			updateReq.Update.AdditionalEnvVars = envVars
-		}
-		updateMaskPaths = append(updateMaskPaths, "additional_env_vars")
-	}
-
-	if !data.SpecsConfigJson.Equal(state.SpecsConfigJson) {
-		if !data.SpecsConfigJson.IsNull() {
-			val := data.SpecsConfigJson.ValueString()
-			updateReq.Update.SpecsConfigJson = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "specs_config_json")
-	}
-
-	if !data.ServiceUrl.Equal(state.ServiceUrl) {
-		if !data.ServiceUrl.IsNull() {
-			val := data.ServiceUrl.ValueString()
-			updateReq.Update.ServiceUrl = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "service_url")
-	}
-
-	if !data.WorkerUrl.Equal(state.WorkerUrl) {
-		if !data.WorkerUrl.IsNull() {
-			val := data.WorkerUrl.ValueString()
-			updateReq.Update.WorkerUrl = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "worker_url")
-	}
-
-	if !data.BranchUrl.Equal(state.BranchUrl) {
-		if !data.BranchUrl.IsNull() {
-			val := data.BranchUrl.ValueString()
-			updateReq.Update.BranchUrl = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "branch_url")
-	}
-
-	if !data.KubeJobNamespace.Equal(state.KubeJobNamespace) {
-		if !data.KubeJobNamespace.IsNull() {
-			val := data.KubeJobNamespace.ValueString()
-			updateReq.Update.KubeJobNamespace = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "kube_job_namespace")
-	}
-
-	if !data.KubeServiceAccountName.Equal(state.KubeServiceAccountName) {
-		if !data.KubeServiceAccountName.IsNull() {
-			val := data.KubeServiceAccountName.ValueString()
-			updateReq.Update.KubeServiceAccountName = &val
-		}
-		updateMaskPaths = append(updateMaskPaths, "kube_service_account_name")
-	}
-
-	if len(updateMaskPaths) > 0 {
-		updateReq.UpdateMask = &fieldmaskpb.FieldMask{
-			Paths: updateMaskPaths,
-		}
-
-		_, err := tc.UpdateEnvironment(ctx, connect.NewRequest(updateReq))
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error Updating Chalk Environment",
-				fmt.Sprintf("Could not update environment: %v", err),
-			)
-			return
-		}
-	}
-
-	if data.Managed.ValueBool() {
-		bc := r.client.NewBuilderClient(ctx)
-
-		_, err := bc.CreateEnvironmentCloudResources(ctx, connect.NewRequest(&serverv1.CreateEnvironmentCloudResourcesRequest{
-			EnvironmentId: data.Id.ValueString(),
-		}))
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error Bootstrapping Chalk Environment Cloud Resources",
-				fmt.Sprintf("Environment was created but cloud resources could not be bootstrapped: %v", err),
-			)
-			return
-		}
-	}
-
-	tflog.Trace(ctx, "updated a chalk_environment resource")
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-func (r *EnvironmentResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var data EnvironmentResourceModel
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Create auth client first
-	if !data.Managed.IsNull() {
-		bc := r.client.NewBuilderClient(ctx)
-
-		_, err := bc.DeleteEnvironmentCloudResources(ctx, connect.NewRequest(&serverv1.DeleteEnvironmentCloudResourcesRequest{
-			EnvironmentId: data.Id.ValueString(),
-		}))
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error Delete Chalk Environment Cloud Resources",
-				fmt.Sprintf("Environment was created but cloud resources could not be deleted: %v", err),
-			)
-			return
-		}
-	}
-
-	// Create team client with token injection interceptor
-	tc := r.client.NewTeamClient(ctx, data.Id.ValueString())
-
-	archiveReq := &serverv1.ArchiveEnvironmentRequest{
-		Id: data.Id.ValueString(),
-	}
-
-	_, err := tc.ArchiveEnvironment(ctx, connect.NewRequest(archiveReq))
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error Archiving Chalk Environment",
-			fmt.Sprintf("Could not archive environment %s: %v", data.Id.ValueString(), err),
-		)
-		return
-	}
-
-	tflog.Trace(ctx, "archived chalk_environment resource")
-}
-
-func (r *EnvironmentResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
-	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
-}
